more stuff

+++ src/camera.c
@@ -6,7 +6,7 @@
 /*   By: joaoteix <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/03 16:35:10 by joaoteix          #+#    #+#             */
-/*   Updated: 2024/04/01 02:49:30 by jcat             ###   ########.fr       */
+/*   Updated: 2024/04/02 00:41:50 by jcat             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ void	cam_init(t_camera *cam)
 	cam->image_width = 600;
 	cam->image_height = cam->image_width / cam->aspect_ratio;
 	cam->focal_length = 1.0;
-	cam->viewport_width = 2 * tan(deg_to_rad(cam->hfov) / 2)
+	cam->viewport_width = 2 * tan(deg_to_rad((float)cam->hfov) / 2.0f)
 		* cam->focal_length;
 	cam->viewport_height = cam->viewport_width
 		/ ((double)(cam->image_width) / cam->image_height);
